rewrite rabbit interface to be simpler and consistent
The two major changes here are that the handlers to onTopic and onRpc
are now extended off amqplib.Message with the only change that the
content field is widened to be a number of possible types from just
Buffer (to support the ContentType header). Secondly, the function
signature for onTopic and onRpc has been simplified such that the
callback is always the last parameter and that it now accepts an
optional middle argument that's an object of options.

Signed-off-by: Matthew Peveler <matt.peveler@gmail.com>